# catalog/product/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ApplicationSuite": {
        "enum": [
          "climate-disaster-resilience",
          "mobility-accessibility-operations",
          "population-health-environmental-exposure",
          "housing-land-use-regeneration",
          "public-service-operations",
          "infrastructure-finance-asset-risk",
          "behavioral-policy-equity"
        ],
        "title": "ApplicationSuite",
        "type": "string"
      },
...
      "CityPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CitySummary"
            },
            "title": "Items",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": [
          "pagination",
          "items"
        ],
        "title": "CityPage",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProductTier": {
        "enum": [
          "G",
          "S",
          "D"
        ],
        "title": "ProductTier",
        "type": "string"
      },
...
      "ScenarioPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScenarioSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": [
          "pagination",
          "items"
        ],
        "title": "ScenarioPage",
        "type": "object"
      },
      "ScenarioStatus": {
        "enum": [
          "completed",
          "screened",
          "insufficient-evidence",
          "infeasible"
        ],
        "title": "ScenarioStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "CivicDecision OS"
    },
    "description": "Read-only projections of versioned public sources, city bundles, 240 evidence-gated scenario designs, scenario screens, DecisionPacks, and analytical benchmarks. Designs remain distinct from executions; the API preserves negative releases and never upgrades the claims in an underlying artifact.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "summary": "Evidence-typed urban decision artifacts",
    "title": "CivicDecision OS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cities": {
      "get": {
        "operationId": "cities_api_v1_cities_get",
        "parameters": [
          {
            "in": "query",
            "name": "tier",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProductTier"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tier"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "country_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2,
                  "minLength": 2,
                  "pattern": "^[A-Za-z]{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country Code"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CityPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cities",
        "tags": [
          "cities"
        ]
      }
    },
    "/api/v1/cities/{city_id}": {
      "get": {
        "operationId": "city_api_v1_cities__city_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "city_id",
            "required": true,
            "schema": {
              "maxLength": 160,
              "minLength": 1,
              "pattern": "^[a-z0-9]+(?:[._-][a-z0-9]+)*$",
              "title": "City Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CityDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "City",
        "tags": [
          "cities"
        ]
      }
    },
    "/api/v1/decision-packs": {
      "get": {
        "operationId": "decision_packs_api_v1_decision_packs_get",
        "parameters": [
          {
            "in": "query",
            "name": "city_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ScenarioStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decision Packs",
        "tags": [
          "decision-packs"
        ]
      }
    },
    "/api/v1/decision-packs/{execution_id}": {
      "get": {
        "operationId": "decision_pack_api_v1_decision_packs__execution_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "maxLength": 160,
              "minLength": 1,
              "pattern": "^[a-z0-9]+(?:[._-][a-z0-9]+)*$",
              "title": "Execution Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionPack"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decision Pack",
        "tags": [
          "decision-packs"
        ]
      }
    },
    "/api/v1/decision-packs/{execution_id}/brief": {
      "get": {
        "operationId": "decision_brief_api_v1_decision_packs__execution_id__brief_get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "maxLength": 160,
              "minLength": 1,
              "pattern": "^[a-z0-9]+(?:[._-][a-z0-9]+)*$",
              "title": "Execution Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "const": "markdown",
              "default": "markdown",
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decision Brief",
        "tags": [
          "decision-packs"
        ]
      }
    },
    "/api/v1/design-families": {
      "get": {
        "operationId": "scenario_families_api_v1_design_families_get",
        "parameters": [
          {
            "in": "query",
            "name": "suite",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ApplicationSuite"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Suite"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioFamilyPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scenario Families",
        "tags": [
          "scenario-designs"
        ]
      }
    },
    "/api/v1/design-families/{family_id}": {
      "get": {
        "operationId": "scenario_family_api_v1_design_families__family_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "maxLength": 160,
              "minLength": 1,
              "pattern": "^[a-z0-9]+(?:[._-][a-z0-9]+)*$",
              "title": "Family Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioFamilyDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scenario Family",
        "tags": [
          "scenario-designs"
        ]
      }
    },
...
    "/api/v1/designs/{design_id}": {
      "get": {
        "operationId": "scenario_design_api_v1_designs__design_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "design_id",
            "required": true,
            "schema": {
              "maxLength": 160,
              "minLength": 1,
              "pattern": "^[a-z0-9]+(?:[._-][a-z0-9]+)*$",
              "title": "Design Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDesignDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scenario Design",
        "tags": [
          "scenario-designs"
        ]
      }
    },
...
    "/api/v1/scenarios/{execution_id}": {
      "get": {
        "operationId": "scenario_api_v1_scenarios__execution_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "maxLength": 160,
              "minLength": 1,
              "pattern": "^[a-z0-9]+(?:[._-][a-z0-9]+)*$",
              "title": "Execution Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scenario",
        "tags": [
          "scenarios"
        ]
      }
    },
...
  }
}